requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
click==5.1
compass-stylesheets==0.12.6
Django==3.0.4
django-filter==2.2.0
django-mustache==1.1.2
djangorestframework==3.11.0
FormEncode==1.3.1
html-json-forms==1.1.1
idna==2.9
linecache2==1.0.0
natural-keys==1.6.0
Pillow==7.0.0
pprint==0.1
psycopg2-binary==2.8.4
py-mini-racer==0.1.18
PyBabeljs==0.0.8
pyScss==1.3.5
pystache==0.5.4
pytz==2019.3
pyxform==1.1.0
PyYAML==5.3
requests==2.23.0
requirejs==0.2.1
six==1.14.0
sqlparse==0.3.1
traceback2==1.4.0
unicodecsv==0.14.1
unittest2==1.1.0
urllib3==1.25.8
wq==1.2.0
wq.app==1.2.0
wq.core==1.2.0
wq.db==1.2.0
wq.start==1.2.0
xlrd==1.2.0
xlsconv==1.1.0
